chore: changed from Integer to Long primary key

+++ src/main/java/br/com/personalgymapi/domain/repository/TrainingRepository.java
@@ -3,5 +3,5 @@ package br.com.personalgymapi.domain.repository;
 import br.com.personalgymapi.domain.entities.Training;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface TrainingRepository extends JpaRepository<Training,Integer> {
+public interface TrainingRepository extends JpaRepository<Training,Long> {
 }
